story #8833: activate ALv2 at import
Import a project with the attribute "use-natures" of
the "trackers" tag set to true.
Nature should be enabled.

Import tracker on an existing project with nature
disabled with the attribute "use-natures" of
the "trackers" tag set to true.
Nature should be enabled.

Import tracker on an existing project with nature
enabled with the attribute "use-natures" of
the "trackers" tag set to false.
Nature should be disabled.

Import tracker an existing project with nature
enabled and no tag. The settings must remain the same.

Import tracker an existing project with nature
disabled and no tag. The settings must remain the same.

Change-Id: I73a28ba7c259903049333a0901c4d72d12cef902