[Development] Deploy our own git hooks
This is part of request #9011: Enforce PSR-2 for Tuleap source code

Allow new comers to follow best practices by enforcing standards.

The first one is to forbid trailing whitespaces.

Run `make dev-setup`.

If you already have a pre-commit file, the hook is not modified and
everything will work as before.
If you don't have pre-commit file, the dev-setup will create one for
you. In that case, you won't be able to commit a file with trailing
whitespaces. For example:

$ echo "   " >> plugins/tracker/include/Tracker/Artifact/Tracker_ArtifactCreator.class.php
$ git commit -a

Next steps:
- add pre-commit hook to enforce PSR-2 standard
- remove existing pre-commit hook to put our own

Change-Id: I8c2cb10fcb14a09e517beb7ab2fa811c10790edc