Display formatted content in custom columns
This is part of story #8940: have configurable columns based on natures

If the format is empty, then the Xref is used.

If the format does not contain any %field, then all artifacts are
displayed the same.

If an artifact cannot be formatted because it does not have every fields
required by the format, then the Xref is used instead with a warning
sign for the end user.

CSV export is not changed (still ids).

Change-Id: I71ff3ceeeee5a3881ad44e0b0a31aa88bfd24704