request #9079 Gerrit REST with BasicAuth
Allow site admin to configure a gerrit server with Basic Auth instead of
Digest.

How to test ?

Modify your gerrit config and set gitBasicAuth = true
in [auth] section.
Be careful, the http password is then the one defined in LDAP (and not the
generated one in gerrit interface)

The you can update the configuration in Tuleap / git site admin

Change-Id: Ie2ea0f44e9aa6ef2af0db8f02ac7fdac523029eb