The resize of a table should occur only on a th/td
Part of story #8940: have configurable columns based on natures

When you click slightly at the right of the cog in a configurable
column, sometimes the library that handle the resize of columns thinks
that you want to resize the inner table (yes we still have tables in
tables still in 2016 ¯\_(ツ)_/¯) and it moves the cog icon to the left.
It may bring some confusion/frustration.

We need to hack the TableKit library to only resize column when we are
on a td/th and not when we are on an icon.

* icon should not move anymore when we click on it
* columns are still resizeable

Change-Id: I5dbdc09c4cc821e69b70b3148adceee4d22ef095