Export formatted column in csv
This is part of story #8940: have configurable columns based on natures

When you export in csv a custom column, the content is formatted
accordingly. If an artifact link cannot be formatted (missing field,
incorrect type of field), then only the id is part of the content.

Aggregates should not be broken.

Escape format since it can contains ' or " and it should not break the
html table nor the popover.

Change-Id: I27ab7347d48e224e2d8c79254f57637ea7cf578d