Import cmmtXXXX references and treat them as links
This task is part of story #9027 Preserve links at import

Install the plugin, then xml import a project with a svn repository that
contains references. Ex:

    <?xml version="1.0" encoding="UTF-8"?>
    <project
     unix-name="project42"
     full-name="Project 42"
     description="Secret project to find the answer"
     access="public">

     <long-description>
      This is the long description of project 42
     </long-description>

     <services>
      <service shortname="plugin_svn" enabled="true" />
     </services>

     <svn>
      <repository name="MyRepository" dump-file="svndump.dump">
       <references>
        <reference source="cmmt2016" target="2" />
       </references>
      </repository>
     </svn>

    </project>

If the svndump contains the revision 2, then when you create a comment in
a changeset with the following body:

  See cmmt2016 or cmmt2017 for details.

Then you can click on cmmt2016 and you are redirected to the revision 2
in the svn repository MyRepository. You cannot click on cmmt2017 as it
is unknown.

If the target revision does not exist, then the link is still extracted
bc the svn revisions are only managed by svn server and Tuleap does not
know if it exists or not. User will get a "404" page.

Change-Id: I55c84da3f218534c6afb3bda35690191af58e2ab
Bonus: remove dead code in svn plugin.