Sass files in Tuleap sources can be built
This is part of story #9132: Migrate from Less to Sass

To test it, you can create a scss file and check that you got
a CSS file.

make sass-dev-docker does not currently watch file modification.
For our Mac OS X friends we have to poll file editions which is
very CPU intensive and seems to raise a lot of errors.
For now, when you build your file in dev mode the generated CSS are not
compressed and sourcemap files are created to ease debugging.

Change-Id: I5bfe862c7069e7d877c5e479929cf954fee796bb