Plugin Bot Mattermost-git - push notification
part of story #9130 Send notification through Mattermost when some Git actions are performed

Dependances :
* BotMattermost Plugin
* Git Plugin

The Plugin is able to choose bot, in git configuration,
to have notification in Mattermost when a push is performed.
New test files for EncoderMessage.php and NotificationMaker.php
and class declaration at the top level.

Change-Id: I4ec83771705a15497e9a7870cafe1e8e14acf055