request #9137 Warning on project import
Try to import import a project with a tracker, bind values and
changesets: it should not output anymore warnings.

When a tracker is imported, we first import the structure, then the
artifacts. At the end of the structure import, caches are cleaned[1]. This
means that when we import an artifact, we must load fields from the
database.

The issue is the following: when we try to import data for a list field,
it tries to validate the corresponding workflow, whereas the workflow
has been previously disabled during the import[2]. The field gets the
workflow from the tracker, but it doesn't know its tracker yet and load
the tracker from the database, thus loosing the workflow disabled state.
Until this commit.

[1] see TrackerXmlImport::createFromXML() (Id60ab8620861096fd1b19a55a7ed8987189bacaa)
[2] see Tracker_Artifact_XMLImport::importBareArtifactsFromXML()

Change-Id: I6fa4262672053ffc9c6fe1d6a79172fe1577dc7e