Add artifact representation into frs_release/{:id}
This is part of story #9087 : see linked artifacts in releases

The route frs_release/{:id} returns the artifact representation
for the artifact ID linked to the release.

If no artifact_id is provided or if the artifact_id does not
exist or if the user is not able to see the artifact,
null is returned instead of the representation.

Change-Id: I2beef76e372d3362274962a0360c358446051d36