request #9145: Notifications are not sent when a Git repository uses hudson_git plugin
The script used to generate the mail expect that the current directory is the Git repo.
When the Hudson Git plugin is activated, curl will be called which we will change the
current directory for the next exec() call. To avoid this kind of magic we set the path
to the Git repository in an environnement variable.

Change-Id: I9bcbbf1d5e54606585a96aa73e0a3b6a9320c718