Display current Jenkins in a dedicated table
This is part of story #9083: trigger webhooks on git push

No functional change here, just introducing a new table to display the
current jenkins hook.

If no jenkins hook configured, then the table presents an empty state
and the list of triggered logs is not displayed.

Next steps (not in that order):
- display list of triggerd logs in a modal
- edit jenkins in a modal
- create jenkins in a modal
- delete jenkins
- display number of triggered jobs in additional column
- start work on generic hooks

Change-Id: I8e3b24f32261ea76aeec04e67ed4c7bbfe4630f1