Open a modal to create/edit Jenkins webhook
This is part of story #9083: trigger webhooks on git push

Introduce a button to create a Jenkins webhook and a link in the table
to edit the webhook. Both open a modal.

If a Jenkins webhook is already configured, then the button is disabled
and a tooltip is displayed on hover.

Please note that the button will be moved on the header line later (see
mockups).

Technical note: Bootstrap does not display tooltips on a disabled button,
that is why there is a wrapper around it.

Next steps (not in that order):
- display list of triggerd logs in a modal
- delete jenkins
- display number of triggered jobs in additional column
- start work on generic hooks

Change-Id: I5e2c03ba9e05e324ef49bc77d31c87f46f4dec27