Refactoring: move all webhook management in git plugin
This is part of story #9083: trigger webhooks on git push

No functional change here, move most of mustache stuff in git plugin so that
now it is plugin git that displays list of webhooks and let hudson_git plugin
add its own stuff. A button [add generic webhook] is added, but does nothing;
its presence is to validate that plugins can add stuff in the middle of git
hooks management.

Change-Id: I3618e5a1b8ab6db9c94846aa6411d00a7989ed58