Display list of custom webhooks
This is part of story #9083: trigger webhooks on git push

The list of custom webhooks that are in the database
(plugin_git_webhook_url) are now displayed.

Please note that you cannot add/edit/delete them through the web
interface.

Since there is no logs yet for webhooks, last push column displays
nothing and the modal for the logs is displayed but empty.

Change-Id: Ia9eb97a796f14fabeb6152c2697541603b4c145d