Refactoring for PHP5.6 - Help
This is part of request #9202 Make Tuleap compatible with PHP 5.6

Update src/www/help source code to deal with PHP 5.6.
To test it, please update your php.conf and set php flags
register_globals and magic_quotes_gpc to off

Help is called when you click on the beautiful [?] link.

Change-Id: I80fbf0c1b3af25246d56ee39a47e956069af9ddf