Change to get previous result
This is a part of bugs #27604 Manage concurrent access

To test: Use this patch: https://gerrit.tuleap.net/#/c/5812/
When we update a test we must have in response of the request
correct previous results. When the score is negative we display
score equal to 0. To verify with one client pass tests 'notrun'
to 'passed' to increase your score restart your tuleap-realtime
server and pass tests 'passed' to 'notrun' to decrease your score.
You must see score equal to 0.

Change-Id: I0e810b702834ef92760f21eba7598f5f5f11acc8