Enable computed field edition
part of story #9090 have computed fields compatible with artifact links v2

Users with write access on the field can disable this autocomputation and set a value
User can switch from set value back to automated computation
When the value is manually set, the automated computation is still show as an info
When the tracker uses computed_fields legacy, the autocomputed field is not updatable.
If the autocomplted field v2 has read-only permission, the field should not be updatable.

Change-Id: I5eba75a7d2790460229bccf80f305db02ced3fad