request #9220: Unable to import big xml project archive
If you try to import a huge project archive, then the import fails
because the libxml is capped. As we are importing a project archive in
CLI, we can safely disable the limitation so that the import goes well.

However this limitation should be kept if the import is done through the
web interface, therefore, even if this is not possible as of today, we
check that we are in CLI to not suffer from a DoS attack the day we open
the project import through the web.

See LIBXML_PARSEHUGE[1] documentation for details.

[1] http://php.net/manual/en/libxml.constants.php

Change-Id: I91d5d078321573c4766d242d285b10a47d4309a0