Replace Fast compute option with an Upgrade button
This is part of story #9090: have computed fields compatible with
artifact links v2

How to test this:

In the tracker administration,
Given a tracker with a computed field that was not set in "fast compute"
mode:
    - when you edit this field, then you should see a button to update
this field. You should no longer see a "Fast compute" checkbox.
    - when you click on this button, the field should be converted to
"fast compute" and when you edit it again, you should no longer see the
button.
    - when you submit the edition by clicking on the usual "Submit",
then the field should not be switched to "fast compute"

Change-Id: I3496cece898eac5e73e73dd11a3430bb3bc05295