Fix unit tests to run both on PHP 5.3 and PHP 5.6
Starting PHP 5.6, the default charset is UTF-8 [1] which broke a test.
We try to force a string to be encoded in ISO-8859-1 but in PHP 5.6+
the default expected encoding before conversion is UTF-8 so we lost some
informations in the process which make the test fail. To solve this issue
we set the encoding expected before conversion and do not rely on PHP default.

Tests must run on both PHP 5.3 and 5.6.

[1] https://secure.php.net/manual/en/ini.core.php#ini.default-charset

Change-Id: I2c9eb3522803b2f186c990437b51c2735328b59e