Hide target field name in Computed fields admin
This is part of story #9090: have computed fields compatible with
artifact links v2

How to test this:

In the tracker administration,
When you add a new computed field, then you should not see a "target
field" text input and when you submit, the new field should have its
field name as "target field" (check in DB)

Given an existing computed field that was set in "fast compute" mode and
given the field's name was the same as the "target field"
    - when you edit it, then you should no longer see a "target field"
text input
    - when you edit it and change its name, then the "target field"
should be updated to the new name (check in DB)

Given an existing computed field, whatever its mode and given the
field's name was NOT the same as the "target field" (e.g. target field
was empty), when you edit it, then you should still see a "target field"
text input

In a following commit, a warning will be displayed to the user so that
they put the correct field name and "target field".

Change-Id: I6eb1ac753861997196af8b652f7c69f60f3527df