Warn when target field is different from field name
This is part of story #9090: have computed fields compatible with
artifact links v2

How to test this:

In the tracker administration,
Given an existing computed field, whatever its mode ("fast compute" or
not) and given the field's name was NOT the same as the "target field"
(e.g. target field was empty), when you edit it, then you will see a
warning to set the "target field" equal to the computed field's name.

This also changes the deprecation warning:
Given an existing computed field, whatever its mode and given the
field's name was NOT the same as the "target field":
    - as siteamdin, when you go to the platform's Tracker administration
page, then this field will be listed in the deprecated fields report
    - as tracker admin, when you go to the Trackers main page, then a
warning to update the deprecated fields will be displayed and the
computed field will be listed in it
    - as tracker admin, when you go to the Tracker's administration
page, then a warning to update the deprecated fields will be displayed

Change-Id: I633b8db9f81f57098d15effda994ea5d80ab7bcc