Add base structure to the output
Part of story #9197: Get rid of Bootstrap in site admin

Now BurningParrot outputs a proper html structure (with <head> and
<body>). You can assert that by looking at the generated source or in
the title of the page.

From a development point of view, BurningParrot as a autoload.php file
so that we don't have a huge stack of require_once at the beginning of
the Theme.

Change-Id: I9175952f2a4357fd8ff33efd171f68d97bdd3c2b