Import tlp (aka Tuleap UI Framework) into Tuleap
Part of story #9197: Get rid of Bootstrap in site admin

The visible change is that in BurningParrot the ui appears more broken
than before :). It is a good sign, meaning that the css is loaded. You
can even inspect a button on the admin page and give it the class
'tlp-button-primary' to see the change.

You can access to the doc of tlp at the following url:

https://tuleap_web_1.tuleap-aio-dev.docker/themes/common/tlp/doc/

(adapt the url to you dev instance)

Change-Id: I9517fe3f362bc67023e61c2315c9b4f5d35f00b8