Use tlp-button in admin area
Part of story #9197: Get rid of Bootstrap in site admin

Go to various parts of the site administration. Buttons that 
were bootstrapified in FlamingParrot are now tlpified in 
BurningParrot. It has no impact if you are using another 
theme than BurningParrot.

A final pass will have to be done in order to choose correct
style for buttons, it is not possible right now since the 
layout is broken.

Change-Id: Iaa59ff6dae569d161d517f2070310b057ab38715