Add Git default permissions for branches and tags
This is part of story #8419 have permissions per tags/branches

With this commit, a git administrator will be able to define
default fine-grained permissions for a project.

These permissions are not taken into account for now.
It's not possible to edit or remove them.

You cannot add the same pattern multiple times.

Change-Id: I1542fad8ade2b043b874ee9be8e5bc54b6d01032