Use FlamingParrot variant for BurningParrot
Part of story #9197: Get rid of Bootstrap in site admin

As site admin, choose a variant for FlamingParrot theme (ex:
DarkPurple). Change the theme and use BurningParrot. Go to site-admin,
ex:
https://tuleap_web_1.tuleap-aio-dev.docker/admin/permission_delegation.php

View the source of the page, the loaded css should be (in case you choose
DarkPurple or Purple):

  <link rel="stylesheet"
    type="text/css" href="/themes/common/tlp/dist/tlp-purple.min.css" />

The variant chooser for BurningParrot will be done later.

Change-Id: Ieb1ca4e4d8ac688dc9cbdd04324f2f85495ea6c6