Dropdown projects
Part of story #9197: Get rid of Bootstrap in site admin

To test: When you click on projects dropdown you must
see projects that you administer and/or you belong.
If you click on the line you are redirected on
project's page and if you click on the cog you are
redirected on the configuration's page.

Change-Id: I4c30e6765628cefc8cb39aae568e1a2a41a020f4