Display alerts in siteadmin
Part of story #9197: Get rid of Bootstrap in site admin

Go to site admin » configuration » project configuration » project
description fields edition[1]. When you add or delete a field, a feedback is
displayed. If when adding field there is an error (missing rank for
example) then there is a redish feedback.

This forced us to move Codendi_HTMLPurifier constants to a dedicated file so
that they are available even if the Codendi_HTMLPurifier has not been loaded
yet: due to autoloading, the file that contains both Codendi_HTMLPurifier class
and constants is loaded only the first time php has to instantiate the class,
we are doomed if we need the constant before the class.

[1] http://tuleap_web_1.tuleap-aio-dev.docker/admin/descfields/desc_fields_edit.php

Change-Id: I2bdb9e211de125fdb62f58847b80167200a6575f