A non migrated field shouldn't be updatable
part of story #9090 have computed fields compatible with artifact links v2

Have a non migrated field, give it update and submission permissions.
Without the patch, the field is updatable in modal V1, Rest Api and mass changes. It is not displayed correctly in edit form.

With the patch the field is displayed but you can never update it. In REST route if you try to update the field
a specific error is catched and displayed.

Done for :
  Modal V1
  Edit Form
  REST API
  Mass changes

Modal V2 - V3 and import/export XML will be in their respective commits

Change-Id: I5dddcca6caf79ccae03e9fde34153f745f5a8a33