request #9312 Replication from gerrit server does not work over http
This commits add a new parameter in Gerrit server configuration
to be able to the replicatation over HTTPS (or HTTP, but please use
HTTPS).

To do this, the Gerrit administrator has to enable the Gerrit
replication plugin. Then, he must configures the replication with the
replication.config file and provide a Tuleap HTTPS URL.

After that, he has to add the forge__user login and the newly provided
password in the secure.config file to add the basic authentication crendentials.

With that, the replication will works through HTTPS.

Be careful, if you use a self signed certificate, you have to import it
in your JAVA keytool[1]

[1] http://alvinalexander.com/java/java-using-keytool-import-certificate-keystore

Change-Id: I797042fe113e18eeb7bbb6359dcd024b9faa0dc8