Fix wrong computed calculation
This is part of story #9090 have computed fields compatible with artifact links v2

This error was due to SQL that returns sometimes values in different
orders. With that, some artifacts were counted twice.

To reproduce this, create three artifacts (named here A, B and C)
in the same tracker that has a computed field.

From artifact A, add artifacts B and C as linked artifact.
Add 3 as manual value for the computed field of artifacts B and C.

The expected value for A must be 6. But sometimes, the value displayed is 9,
which is wrong.

Change-Id: I9c920028b42f5ae691992ef268c7b3b2caf47cd9