Add a timeout to expensive operations in the post-receive-email Git hook
I have voluntary only targeted the most expensive operations to avoid
sending an email completely meaningless. If we see with real data this is not enough,
we can easily do something more aggressive.

This is part of story #9213: timeout git operations in push email

You can test the patch by pushing a large number of commits at once or with
a lot of files or with very large files.

Change-Id: I65c91da30209bf6f929b86f3fa6184194727af1b