Fix creation of children in cardwall
This is part of story #9090 have computed fields compatible with
artifact links v2

How to test this:
Given a User story tracker with a _required_ computed field and given a
hierarchy as follows:

User Story
`-- Tasks

Given an existing User story US in a Cardwall (e.g. a Sprint cardwall),
When I create a new Task that should be linked to that User Story
Then the task should be created and correctly linked to US

Without this patch, the task will be created but will have a silent
error and won't be linked to US because of the required computed field.

This fix also has an impact on copy artifact
please note like other field, when no value is provided the field is set to default value (autocomputed)

Change-Id: I12bf8b067ade509364b479d4f608ae635273257a