request #9319: Update artifact-modal version
How to test this:

In www/scripts/angular, please run 'bower install && npm run build'

Add a string field to the Validation Test Definition tracker. Set its
permission: nobody should be able to submit it.

When you try to add a new test (to any environment), the string field
should not be displayed. Without this patch, it was displayed.

Change-Id: Ia0fde23355c941d5724111daa1afddef26813c78