request #9336 Computed calculation is broken for old projects
Given a tracker A with a computed field which as a computed field total effort and who has a target field name effort estimate and which is checked as Fast compute.
Given an other tracker B with a computed field named total effort for a target field effort_estimate,
the calcul must display the correct value.

Change-Id: I51d2cdb8e26480fcb8f8e378d6058515b48eea18