Import/Export of artifactlinks
request #9010 Export project tool must export all importable services

You should be able to export and reimport a project without custom nature.
In admin, artifact link nature the project is not in list of projects which can use artifact links nature.
In trackers, the hierarchy is kept.

You should be able to export and reimport a project with custom nature.
In admin, artifact link nature the project is in list of projects which can use artifact links nature.
In trackers the nature type is kept.
If a nature doesn't exist on platedform it is created with its shorname for label and reverse label.

We should also be able to unlink artifacts with nature (change its father)
We should also be able to change nature to no nature of an artifact

Change-Id: I6aac622872945db87e668f1b552cde7d346d612c