Replace angular-ui-tree with dragular in kanban view
This is part of story #8808: Handle collapsed columns

How to test this:
- Run 'make sass-docker' in the tuleap dir
- Run 'bower prune && bower install && npm run build' in the kanban dir

- Drag and drop items from/to the backlog, the archive, another column
- Reorder items in the backlog, archive, another column
- Drop an item in an empty open column. The image should change color
according to your theme
- Start dragging an item, and press the <kbd>esc</kbd> key. It should
cancel, the item should go back to its starting place.
- Drop an item in a closed column. The background of the column should
change according to your theme. It should work when the column has been
loaded and when it has not (if you close a column and reload the page,
the column will stay closed and won't be loaded)

- Using the edition artifact-modal, change the status of a card. It
should move at the end of the corresponding column.
- Filter items. The filter should still work
- When there is an error with the REST move / reorder, the error modal
should still be displayed to the user

Change-Id: Id33e431a3100094c175b65d016c7a984671920d0