request #8839: Call to sparklines.php should be made in POST to avoid errors
Currently, when you have a lot of references you will get an HTTP error 414
"Request URI too long".
This cause two issues:
 * sparklines can not be processed as expected since you never get
the expected answer from the server
 * it makes the error logs of the web server less useful since we have a lot
of garbage data

RFC 7231 in section 4.3.3 [1] says that POST can be used for
"Providing a block of data [...] to a data-handling process" and sparklines.php
can be qualified of a data-handling process.

Please note, the same patch has been proposed in commit 60cf808 and then reverted
in 8cad956 but I can not find a reason for the initial change to be reverted aside
from the fact the initial contribution was published during a merge with another
patch that is not related to it.

[1] https://tools.ietf.org/html/rfc7231#section-4.3.3

Change-Id: I3b4fcefd63307b2257cadf8f1a994c5344fb31be