Define a hierarchy with the artifact links v2
This is part of story #9305 have agile dashboard compatible with ALv2

In a project that already uses the artifact link v2,
you are now able to define a Tracker hierarchy.
Once done, this hierarchy will define the usage of the
"_is_child" nature.

For instance, if I've got the following hierarchy:

User story
`
`_Tasks

If I link a task to a user story, this link will automatically
be qualified as "_is_child", whatever the type manually choosen.

If I link a requirement to a user story with the type
"_is_child" manually set, the link will be qualified as "no type"

Without hierarchy defined, you can set as you want links with
the "_is_child" type

Triggers and Burndown will be reactivated in a future commit.

Change-Id: Id878220537c5cd9f28369855783573f43813764c