Agile Dashboard can be used with artifact links v2
This is part of story #9305 have agile dashboard compatible with ALv2

In a project that uses the artifact links v2, a project administrator should be able to
activate the Agile Dashboard service. Once done, this service could be
fully usable (in the same way regarding projects without artifact links
v2).

Change-Id: I262f21a91292029c263bd320c5ff953cfc3d8bba