story #9251: mass remove links when editing artifacts
How to test this:

When you edit an artifact link field, you can now click on a checkbox in
the header of the table to mark all links of the current pane (e.g.
Tasks) for deletion.
When you add temporary links (using the '+ Add artifacts' button), those
links can also be removed using the checkbox

Change-Id: I13838bd469ddd0f81208cf04fa5c1e9d15657a90