Do not raise perms error when the transition is not valid
Part of story #9096: Define comment as mandatory in a tracker workflow transition

When an artifact is updated, and the condition for the transition is not
respected, then the feedback always says that this is because
permissions are not respected, which is not always right.

On a transition with mandatory comment, you should not have a "perms"
error if the comment is empty while updating the artifact.

You should still have the "perms" error if you don't have the right to
pass the transition though.

(This bug is not present in modal v2)

Change-Id: I51e66b45574eaded721a5617303b155980966c8a