Create references pr and pullrequest
This is part of  activity #28334 Cross references (back to pullrequests)

The plugin pullrequest deploy 2 system references: pr and pullrequest.
You can now reference a pullrequest by its id in an artifact comment for
instance. Once done, you will be able to click on the link and go to the
pullrequest view.

If a project reference named pr or pullrequest was manually created
before this patch, the new system reference will not be taken into
account in this project.

The cross reference part will be done in a future commit.

Change-Id: Ic3d928a55c2c95bacdc8d56056a82bd3dae96e84