HTML tags are no more displayed in edition form and in selectbox label
This is part of activity #28334 Cross references (back to pullrequests)

Before this patch, if the title of your pullrequest contains a reference (i.e art #8),
the title in the selectbox will be like '... <a href ... art #8 </a> ..."
and the content in the edit form will have this HTML entities displayed.

Now, these HTML entites are no longer displayed.

Change-Id: If4f7a586be9d9b6ff57964af57f4f3896a321b49