SVN custom logs are not working in new installation
This is part of story #9393 See svn plugin access logs

In new tuleap installation (with tuleap-vhost.conf apache file),
the SVN custom logs are not working anymore.

There was a change in the inclusion of *.conf file outside vhost,
so this CustomLog is not taken into account.

To test it (if you are using the new tuleap apache configuration),
add by hand in tuleap-vhost.conf the following line at the end of the
first <VirtualHost> (line 20):

CustomLog logs/svn_log "%h %l %u %t %U %>s \"%{SVN-ACTION}e\"" env=SVN-ACTION

Change-Id: I2e1210fa543bb5ec0a74da99f2fd0b7b74fd88d9