Create db_stats_svn_plugin_history.pl script
This is part of story #9393 See svn plugin access logs

This patch introduces the perl script with the logical
part but not yet the save in Database.

How to test it ?
----------------

* Do some SVN operation from core and plugin in several projet with
several users. Be careful, if you are using the new apache
configuration, please take a look at the following Git commit:
c2f400c3f974f60f911c004d9644c790bf54462d

* Save the content of /var/log/httpd/svn_log into
/var/log/tuleap/2016/08/svn_20160823.log. Adapt the date if
necessary (this is the exepected filename if you test this patch today)

* In db_stats_svn_plugin_history.pl, uncomment the line 159

* As root, launch /usr/share/tuleap/src/utils/compute_all_daily_stats.sh
and check if the output is correct.

Change-Id: I3568f3831b4964670064364698066fad176d244c