request #9420: Openlist field can have an erroneous value when an artifact is created
To be able to see the bug, you should use an open list field without a default value.

You can easily reproduce the bug by using a kanban based on a tracker with an
open list field. With this contribution, after the card is created you should not
have the user None in the list.

Change-Id: I2f6ff236ad2bc3f91884afa8af6ffd67a707a9b7