Have consistent selectbox for FRS release
story #9409 define FRS permissions for packages and releases

At release creation, proposed permissions are already OK as they
are inherited from package.

This patch corrects permissions for already existing release.

Please note that permissions should be updated when you change
global access or public project visibility : if a release has
a permission which is not possible anymore, the permission
should be erased.

Change-Id: Ideff7f1d00f1e106e59d94b77a350ecbc150a2b4