Use viewvc-theme-tuleap
This is related to story #8597: have a recent version of ViewVC

First this commit should not break display of SVN plugin, nor legacy
viewvc in SVN plugin, SVN core, and CVS.

Then this commit improve display only for new ViewVC in SVN plugin. SVN
core and CVS will be done in a next contribution (they are probably
broken with this patch).

If you test new display of ViewVC, you must:
1. install ViewVC from epel repositories
2. install viewvc-theme-tuleap that you can build from corresponding
pull request[1] on tuleap.net.

Furthermore we now always return an image for the users avatar. Before
that, when we were requesting the avatar for a non existing user
(/users/jdoe/avatar.png) the resulting page was html stuff. This causes
issue in browser that are expecting image/* MIME type. Now we don't have
anymore the broken image icon on the avatar.

[1] https://tuleap.net/plugins/git/?action=pull-requests&repo_id=222&group_id=101#/pull-requests/17/overview

Change-Id: I782eb229196f19565c4fab401640bd49293b901a