request #9449: Remove Gravatar usage
It does not appears we have a lot of cases where an avatar is requested
for an anonymous user (I'm not able to find one) under normal usage.

So this is either dead code or Tuleap depending on a external service, these
two reasons are valid to remove the code.

Change-Id: I1a6cd0d53cb28a47120f4c474a1f9a730f529d8b