request #9452: CSRF check error when trying to restore a tracker
How to test this:

Without this patch, edit CSRFSynchronizerToken.php and set
MAX_TOKEN_PER_STORAGE to 10 (for example).
Given you have deleted more than 10 trackers.

As siteadmin, go to Site administration > Configuration (in the center)
> Trackers Administration > Pending trackers removal
There, click on "Restore" for the first tracker.

Without this patch, an error "Your requested does not appear to be
valid.  Please retry." will be displayed. With this patch, the tracker
will be correctly restored.

Change-Id: I9c32d870ca06285088f82e64a507a65c1e97f411