request #9578: Session tokens are stored in clear in the database
The goal of this contribution is to:
 * limit the impact of a leak of the session table
 * prevent timing attack existing with the current implementation

In order to do that we hash the session token before storing it. Unlike
passwords, we use a fast cryptographic hash function to limit the impact
on performance. Also, the token is a large cryptographic random value so
even with a fast hash function the bruteforce is not easy. You basically
have to bruteforce the whole keyspace of the random value or find a collision
in the hash function.
Since we are not able to compare the token provided by the user and the one
stored in constant time using the database engine, we need something to select
the proper session before checking the token. This implies that all existing
sessions will be invalidated by this change.

Change-Id: Ic3ee00c9f4277aae57aa4f41f85dd201da61bac7