request #9511: Unable to add an artifact link via REST
If you try to add an artifact link via the rest api, you get an error:

  {
    "error": {
      "code": 500,
      "message": "Internal Server Error: error: Nature is missing
	  for artifact #2514.\n"
    }
  }

There is no need to check for nature. If no nature is given then there
will be no nature for the link ¯\_(ツ)_/¯.

It should not break Web UI, CSV import, REST, Soap, DCOM, Corba.

Change-Id: I8b8086b1f5ebf3f597129f034fa36248172c2430