request #9518: html entities displayed in kanban cards
The title of an artifact should not output html entities.

To reproduce, bind the semantic `title` to a text field,
submit an artifact with the title containing some html content (use
bold, italic, marquee, whatever). The title of the artifact at the top
of the page should not display markup. This fix issue also:

* on the card in the kanban,
* on my personal page (My Artifacts widget)
* on agile dashboard homepage, content and planning view
* artifact modal v3 in a dedicated review: https://gerrit.tuleap.net/6544 (I4ed07dc632aee3d28860f18f3761233a0bf0fcec)

Change-Id: I75b076a9d3292c9fb5c09a31cdc335d5e854379e