Remove rogue artifacts
Part of request #9520: Rogue artifacts created when global rules are violated

In an empty tracker, create global rules. Try to submit an artifact that
violates those rules. You got an error message but an entry is created
in the database (artifact with last_changeset_id = 0). You can assert
this by going to the tracker service homepage, the total number of
artifact for the empty tracker is 1 instead of 0.

This forgeupgrade cleans the database and remove those artifacts (and
only them, hopefully).

Change-Id: I9d0c3bb0def38bba5dbf37201c6c66bca7de5769