Prevent creation of rogue artifacts
Part of request #9520: Rogue artifacts created when global rules are violated

In an empty tracker, create global rules. Try to submit an artifact that
violates those rules. You got an error message but NO entry is created
in the database. On the tracker service homepage, the total number of
artifact for the empty tracker is still 0 (and not anymore 1).

Change-Id: If5b1a4398f8350ebea8caba3c2c593b8b09883d7